Convert files to ts and export executable start fn
