+++ packages/server/src/start.ts
@@ -2,10 +2,14 @@ import express from 'express'
 
 import server from './index'
 
-const app = express()
+const start = () => {
+  const app = express()
 
-app.use(server)
+  app.use(server)
 
-app.listen(3000, () => {
-  console.log('Server listening on port 3000')
-})
+  app.listen(3000, () => {
+    console.log('Server listening on port 3000')
+  })
+}
+
+export default start
